requirements.txt
albumentations==1.3.1
comet_ml==3.35.1
coremltools==7.0
easydict==1.10
export==0.2.1
Flask==3.0.0
GitPython==3.1.38
GitPython==3.1.40
ipython==8.16.1
matplotlib==3.7.3
matplotlib==3.8.0
mss==9.0.1
numpy==1.25.2
onnxruntime==1.16.1
opencv_python==4.8.1.78
openvino==2023.1.0
paddle==1.0.2
pafy==0.5.5
pandas==2.1.2
Pillow==10.0.1
Pillow==10.1.0
psutil==5.9.0
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
scipy==1.11.3
seaborn==0.13.0
setuptools==68.0.0
tensorflow==2.14.0
tensorrt==8.6.1.post1
tflite_runtime==2.14.0
thop==0.1.1.post2209072238
torch==2.1.0+cu118
torchvision==0.16.0+cu118
tqdm==4.66.1
train==0.0.5
tritonclient==2.38.0
ultralytics==8.0.199
